add parsing error variant and organize code as a library
